Implemented toggle command
