+++ src/main/java/me/dueris/genesismc/core/DataContainer.java
@@ -1,5 +1,6 @@
 package me.dueris.genesismc.core;
 
+import me.dueris.genesismc.core.commands.ToggleCommand;
 import me.dueris.genesismc.core.utils.ShulkUtils;
 import org.bukkit.NamespacedKey;
 import org.bukkit.entity.Player;
@@ -34,6 +35,9 @@ public class DataContainer implements Listener {
     }
     if (!p.getPersistentDataContainer().has(new NamespacedKey(GenesisMC.getPlugin(), "in-phantomform"), PersistentDataType.INTEGER)) {
       p.getPersistentDataContainer().set(new NamespacedKey(GenesisMC.getPlugin(), "in-phantomform"), PersistentDataType.INTEGER, 1);
+    }
+    if (!p.getPersistentDataContainer().has(new NamespacedKey(GenesisMC.getPlugin(), "toggle"), PersistentDataType.INTEGER)) {
+      p.getPersistentDataContainer().set(new NamespacedKey(GenesisMC.getPlugin(), "toggle"), PersistentDataType.INTEGER, 1);
     }
 
   }
